requirements.txt
APScheduler==3.11.0
asgiref==3.9.1
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.5
django-cors-headers==4.7.0
django-dbml==0.10.1
django-rest-swagger==2.2.0
djangorestframework==3.16.1
docopt==0.6.2
docx2pdf==0.1.8
drf-yasg==1.21.10
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
lxml==6.0.1
MarkupSafe==3.0.2
num2words==0.5.14
numpy>=2.0.0,<3.0.0
openapi-codec==1.3.2
packaging==25.0
pandas==2.3.2
promise==2.3
psycopg==3.2.9
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
# pywin32==311
PyYAML==6.0.2
requests==2.32.5
simplejson==3.20.1
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0